Remove unused classes and add new model classes for a game
This commit removes some unused classes that were there for testing purposes and adds several classes that represent the building blocks for a game. These new classes include the base classes for activating, using, and managing the game entity, items, actors, room effects, charges, etc. This provides a new structure to the project which will help in game interactions and simulations.